requirements.txt
boto3>=1.26.0
pillow>=9.0.0
tqdm>=4.64.0
tabulate>=0.9.0
requests>=2.28.0 